refactor(maestro): split ide/diagnostics into cohesive modules
