+++ crates/vize_maestro/src/ide/diagnostics.rs
@@ -7,15 +7,25 @@
 //! - Future: vize_canon (type checker)
 #![allow(clippy::disallowed_types, clippy::disallowed_methods)]
 
+mod builder;
 mod collectors;
 #[cfg(feature = "native")]
 mod corsa;
+mod line_index;
+mod service;
+mod severity;
+#[cfg(test)]
+mod tests;
 
-use tower_lsp::lsp_types::{Diagnostic, DiagnosticSeverity, NumberOrString, Range, Url};
+pub use builder::DiagnosticBuilder;
+pub use service::DiagnosticService;
+pub use severity::Severity;
 
-use crate::ide::ecosystem;
-use crate::server::ServerState;
-use crate::utils::is_standalone_html_path;
+pub(in crate::ide) use line_index::LineIndex;
+#[cfg(test)]
+pub(in crate::ide) use line_index::offset_to_line_col;
+#[cfg(feature = "native")]
+pub(in crate::ide) use service::{SourceMapping, VirtualTsResult};
 
 /// Diagnostic source identifiers.
 pub mod sources {
@@ -26,1181 +36,4 @@ pub mod sources {
     pub const LINTER: &str = "vize/lint";
     pub const TYPE_CHECKER: &str = "vize/types";
     pub const MUSEA: &str = "vize/musea";
-}
-
-/// Diagnostic severity levels.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Severity {
-    Error,
-    Warning,
-    Information,
-    Hint,
-}
-
-impl From<Severity> for DiagnosticSeverity {
-    fn from(s: Severity) -> Self {
-        match s {
-            Severity::Error => DiagnosticSeverity::ERROR,
-            Severity::Warning => DiagnosticSeverity::WARNING,
-            Severity::Information => DiagnosticSeverity::INFORMATION,
-            Severity::Hint => DiagnosticSeverity::HINT,
-        }
-    }
-}
-
-/// Source position mapping from @vize-map comments.
-#[cfg(feature = "native")]
-#[derive(Debug, Clone)]
-pub(super) struct SourceMapping {
-    /// Byte offset start in SFC
-    pub(super) start: u32,
-    /// Byte offset end in SFC
-    pub(super) end: u32,
-}
-
-/// Virtual TypeScript generation result with position mapping info.
-#[cfg(feature = "native")]
-pub(super) struct VirtualTsResult {
-    /// Generated TypeScript code (post `.vue` → `.vue.ts` import rewrite).
-    pub(super) code: String,
-    /// Byte-range source mappings from generated TS back to the source SFC.
-    /// Offsets are in pre-rewrite generated TS coordinates; callers must
-    /// translate post-rewrite byte offsets via `import_source_map` first.
-    pub(super) source_mappings: Vec<vize_canon::virtual_ts::VizeMapping>,
-    /// Byte-offset mapping from post-rewrite to pre-rewrite virtual TS.
-    /// Empty when no `.vue` import specifiers were rewritten.
-    pub(super) import_source_map: vize_canon::ImportSourceMap,
-    /// Relative `.vue` import specifiers found in the SFC's script. The
-    /// editor session overlays each sibling's virtual TS so relative
-    /// imports resolve under the temp-dir Corsa session (issue #752).
-    pub(super) relative_vue_imports: Vec<std::string::String>,
-    /// Line number where user code starts in virtual TS (0-indexed)
-    pub(super) user_code_start_line: u32,
-    /// Line number where script starts in original SFC (1-indexed)
-    pub(super) sfc_script_start_line: u32,
-    /// Line number where template scope starts in virtual TS (0-indexed)
-    pub(super) template_scope_start_line: u32,
-    /// Line-to-source mappings from @vize-map comments
-    /// Index is virtual TS line number (0-indexed), value is source position in SFC
-    pub(super) line_mappings: Vec<Option<SourceMapping>>,
-    /// Number of import lines skipped from user code (to adjust line mapping)
-    pub(super) skipped_import_lines: u32,
-}
-
-/// Diagnostic service for collecting and aggregating diagnostics.
-pub struct DiagnosticService;
-
-impl DiagnosticService {
-    /// Collect all diagnostics for a document.
-    pub fn collect(state: &ServerState, uri: &Url) -> Vec<Diagnostic> {
-        let Some(doc) = state.documents.get(uri) else {
-            tracing::warn!("collect: document not found for {}", uri);
-            return vec![];
-        };
-
-        let content = doc.text();
-        let mut diagnostics = Vec::new();
-        let features = state.lsp_features();
-
-        if !features.has_diagnostics() {
-            return diagnostics;
-        }
-
-        // Build the line index once for this document. Every collector below
-        // maps byte offsets in `content` to (line, utf16_col) against it, so
-        // sharing one index avoids re-scanning the whole document per offset.
-        let line_index = LineIndex::new(&content);
-
-        // Check if this is an Art file (*.art.vue)
-        let path = uri.path();
-        if path.ends_with(".art.vue") {
-            // Musea-specific diagnostics for Art files
-            if features.lint {
-                diagnostics.extend(Self::collect_musea_diagnostics(uri, &content, &line_index));
-            }
-            // Don't return early here; async collection still adds Corsa diagnostics.
-            return diagnostics;
-        }
-
-        if is_standalone_html_path(path) {
-            if features.lint {
-                let linter_config = state.get_linter_config();
-                let lint_diags = Self::collect_lint_diagnostics(
-                    uri,
-                    &content,
-                    features.ecosystem,
-                    &linter_config,
-                    &line_index,
-                );
-                tracing::info!(
-                    "collect: standalone HTML patina lint diagnostics: {}",
-                    lint_diags.len()
-                );
-                diagnostics.extend(lint_diags);
-            }
-            return diagnostics;
-        }
-
-        // Standard SFC processing — parse once and share the descriptor with
-        // every block-level collector below.
-        let descriptor = match Self::parse_sfc_for_collect(uri, &content) {
-            Ok(descriptor) => descriptor,
-            Err(parse_diagnostic) => {
-                tracing::info!("collect: skipping dependent diagnostics after SFC parse error");
-                diagnostics.push(parse_diagnostic);
-                return diagnostics;
-            }
-        };
-
-        // Collect parser diagnostics for script and template blocks before
-        // dependent analyzers, so broken blocks do not fan out into noisy
-        // lint/type/Corsa diagnostics.
-        let script_diags =
-            Self::collect_script_diagnostics(uri, &content, &descriptor, &line_index);
-        let has_script_parse_error = !script_diags.is_empty();
-        tracing::info!("collect: script parser diagnostics: {}", script_diags.len());
-        diagnostics.extend(script_diags);
-
-        let template_diags =
-            Self::collect_template_diagnostics(uri, &content, &descriptor, &line_index);
-        let has_template_parse_error = !template_diags.is_empty();
-        tracing::info!(
-            "collect: template parser diagnostics: {}",
-            template_diags.len()
-        );
-        diagnostics.extend(template_diags);
-        if has_script_parse_error || has_template_parse_error {
-            tracing::info!("collect: skipping dependent diagnostics after block parse error");
-            return diagnostics;
-        }
-
-        // Surface Vue-specific compile errors (e.g. DEFINE_PROPS_DESTRUCTURE_DEFAULT_TYPE)
-        // that the TypeScript checker cannot derive on its own. Mirrors the
-        // canon path used by `vize check` so editor and CLI stay aligned.
-        let sfc_compile_diags =
-            Self::collect_sfc_compile_diagnostics(uri, &content, &descriptor, &line_index);
-        tracing::info!(
-            "collect: sfc compile diagnostics: {}",
-            sfc_compile_diags.len()
-        );
-        diagnostics.extend(sfc_compile_diags);
-
-        if features.lint {
-            // Collect linter diagnostics (vize_patina)
-            let linter_config = state.get_linter_config();
-            let lint_diags = Self::collect_lint_diagnostics(
-                uri,
-                &content,
-                features.ecosystem,
-                &linter_config,
-                &line_index,
-            );
-            tracing::info!("collect: patina lint diagnostics: {}", lint_diags.len());
-            diagnostics.extend(lint_diags);
-            if features.cross_file {
-                // Cross-file analysis is opt-in (defaults to off) because it
-                // touches every Vue file in the workspace. When enabled, the
-                // same analyzer groups used by `vize lint --cross-file` join
-                // the editor's diagnostic stream. The actual cross-file
-                // analyzer wiring is being added incrementally — for now
-                // the gate is observable through tracing so callers can
-                // verify the config knob took effect.
-                tracing::info!(
-                    "collect: cross-file lint enabled (groups will surface as they are wired up)"
-                );
-            }
-        } else {
-            tracing::info!("collect: patina lint diagnostics skipped (disabled by config)");
-        }
-
-        if features.ecosystem {
-            let ecosystem_diags = ecosystem::diagnostics(&content, uri);
-            tracing::info!(
-                "collect: ecosystem editor diagnostics: {}",
-                ecosystem_diags.len()
-            );
-            diagnostics.extend(ecosystem_diags);
-        }
-
-        if state.is_lsp_typecheck_enabled() {
-            // Collect type checker diagnostics (vize_canon)
-            let type_diags = super::TypeService::collect_diagnostics(state, uri);
-            tracing::info!("collect: type checker diagnostics: {}", type_diags.len());
-            diagnostics.extend(type_diags);
-        } else {
-            tracing::info!("collect: type checker diagnostics skipped (disabled by config)");
-        }
-
-        // Also lint inline <art> blocks in regular .vue files
-        if features.lint {
-            let inline_art_diags =
-                Self::collect_inline_art_diagnostics(uri, &content, &descriptor, &line_index);
-            tracing::info!(
-                "collect: inline art diagnostics: {}",
-                inline_art_diags.len()
-            );
-            diagnostics.extend(inline_art_diags);
-        }
-
-        diagnostics
-    }
-
-    /// Collect only the lint-sourced diagnostics (`vize/lint`, `vize/musea`)
-    /// for a document.
-    ///
-    /// This is a fast subset of [`Self::collect`] used by the hover handler,
-    /// which only ever reads diagnostics whose source is `vize/lint` or
-    /// `vize/musea`. It reproduces exactly the lint/musea diagnostics that the
-    /// full pipeline would publish — including the parser-error short-circuit
-    /// that gates them — while skipping the expensive SFC compile, ecosystem,
-    /// and (per-call, uncached) SFC type-check passes that hover discards. The
-    /// returned set is byte-for-byte the lint/musea subset of `collect`'s
-    /// output for every document shape (Art file, standalone HTML, SFC).
-    pub fn collect_lint_only(state: &ServerState, uri: &Url) -> Vec<Diagnostic> {
-        let Some(doc) = state.documents.get(uri) else {
-            return vec![];
-        };
-
-        let content = doc.text();
-        let features = state.lsp_features();
-        let mut diagnostics = Vec::new();
-
-        if !features.has_diagnostics() || !features.lint {
-            return diagnostics;
-        }
-
-        // Build the line index once for this document, shared by every
-        // collector below (mirrors `collect`).
-        let line_index = LineIndex::new(&content);
-
-        // Art files (*.art.vue): Musea-specific lint only.
-        let path = uri.path();
-        if path.ends_with(".art.vue") {
-            diagnostics.extend(Self::collect_musea_diagnostics(uri, &content, &line_index));
-            return diagnostics;
-        }
-
-        // Standalone HTML: patina lint only.
-        if is_standalone_html_path(path) {
-            let linter_config = state.get_linter_config();
-            diagnostics.extend(Self::collect_lint_diagnostics(
-                uri,
-                &content,
-                features.ecosystem,
-                &linter_config,
-                &line_index,
-            ));
-            return diagnostics;
-        }
-
-        // Standard SFC: parse once, then mirror `collect`'s parser-error
-        // short-circuit so lint only surfaces when the full pipeline would
-        // also surface it.
-        let Ok(descriptor) = Self::parse_sfc_for_collect(uri, &content) else {
-            return diagnostics;
-        };
-        let has_block_parse_error =
-            !Self::collect_script_diagnostics(uri, &content, &descriptor, &line_index).is_empty()
-                || !Self::collect_template_diagnostics(uri, &content, &descriptor, &line_index)
-                    .is_empty();
-        if has_block_parse_error {
-            return diagnostics;
-        }
-
-        let linter_config = state.get_linter_config();
-        diagnostics.extend(Self::collect_lint_diagnostics(
-            uri,
-            &content,
-            features.ecosystem,
-            &linter_config,
-            &line_index,
-        ));
-        diagnostics.extend(Self::collect_inline_art_diagnostics(
-            uri,
-            &content,
-            &descriptor,
-            &line_index,
-        ));
-        diagnostics
-    }
-
-    /// Collect diagnostics asynchronously (includes Corsa diagnostics when available).
-    #[cfg(feature = "native")]
-    pub async fn collect_async(state: &ServerState, uri: &Url) -> Vec<Diagnostic> {
-        tracing::info!("collect_async: {}", uri);
-
-        // Start with sync diagnostics (patina, etc.)
-        let mut diagnostics = Self::collect(state, uri);
-        tracing::info!("sync diagnostics count: {}", diagnostics.len());
-        if has_blocking_parser_error(&diagnostics) {
-            tracing::info!("collect_async: Corsa diagnostics skipped after parser error");
-            return diagnostics;
-        }
-        if is_standalone_html_path(uri.path()) {
-            tracing::info!("collect_async: Corsa diagnostics skipped for standalone HTML");
-            return diagnostics;
-        }
-
-        if state.is_lsp_typecheck_enabled() {
-            // Try to get Corsa diagnostics (with timeout, skip on failure).
-            // Use 10s timeout - polling for diagnostics internally uses 5s
-            let corsa_future = Self::collect_corsa_diagnostics(state, uri);
-            match crate::runtime::timeout(std::time::Duration::from_secs(10), corsa_future).await {
-                Ok(corsa_diags) => {
-                    tracing::info!("corsa diagnostics count: {}", corsa_diags.len());
-                    diagnostics.extend(corsa_diags);
-                }
-                Err(_) => {
-                    tracing::warn!("corsa diagnostics timed out for {}", uri);
-                }
-            }
-
-            // When the user opted into typecheck but Corsa never came up
-            // (init failed, timed out, or simply not yet attempted while we
-            // already produced zero corsa diagnostics for an SFC), surface a
-            // hint diagnostic so the Problems panel reflects what is
-            // happening. Without this, the editor goes silent and users
-            // assume their project is clean. See #681.
-            if !state.has_corsa_bridge()
-                && !diagnostics
-                    .iter()
-                    .any(|d| d.source.as_deref() == Some(sources::TYPE_CHECKER))
-            {
-                diagnostics.push(typecheck_unavailable_hint());
-            }
-        } else {
-            tracing::info!("collect_async: Corsa diagnostics skipped (disabled by config)");
-        }
-
-        diagnostics
-    }
-
-    /// Create a diagnostic from a custom error.
-    pub fn create_diagnostic(
-        range: Range,
-        severity: Severity,
-        source: &str,
-        code: Option<i32>,
-        message: String,
-    ) -> Diagnostic {
-        Diagnostic {
-            range,
-            severity: Some(severity.into()),
-            code: code.map(NumberOrString::Number),
-            source: Some(source.to_string()),
-            message,
-            ..Default::default()
-        }
-    }
-}
-
-/// Build the hint diagnostic surfaced when LSP type checking is requested
-/// but the Corsa bridge is not available. Single point of truth so the
-/// wording stays consistent for tests and follow-up code-action work.
-#[cfg(feature = "native")]
-fn typecheck_unavailable_hint() -> Diagnostic {
-    Diagnostic {
-        range: Range {
-            start: tower_lsp::lsp_types::Position {
-                line: 0,
-                character: 0,
-            },
-            end: tower_lsp::lsp_types::Position {
-                line: 0,
-                character: 0,
-            },
-        },
-        severity: Some(DiagnosticSeverity::HINT),
-        code: Some(NumberOrString::String("typecheck-unavailable".to_string())),
-        source: Some(sources::TYPE_CHECKER.to_string()),
-        message: "Type checking is unavailable in this workspace. \
-            Make sure `tsconfig.json` exists and the Corsa runtime is reachable; \
-            see https://vizejs.dev/guide/static-analysis."
-            .to_string(),
-        ..Default::default()
-    }
-}
-
-#[cfg(feature = "native")]
-fn has_blocking_parser_error(diagnostics: &[Diagnostic]) -> bool {
-    diagnostics.iter().any(|diagnostic| {
-        matches!(
-            diagnostic.source.as_deref(),
-            // Parser-level errors plus SFC compile-time validation errors
-            // both leave the script body in a state where Corsa would just
-            // cascade — the user already has the actionable diagnostic.
-            Some(
-                sources::SFC_PARSER
-                    | sources::SCRIPT_PARSER
-                    | sources::TEMPLATE_PARSER
-                    | sources::SFC_COMPILER
-            )
-        ) && diagnostic.severity == Some(DiagnosticSeverity::ERROR)
-    })
-}
-
-/// Builder for creating diagnostics.
-pub struct DiagnosticBuilder {
-    range: Range,
-    severity: Severity,
-    source: String,
-    code: Option<i32>,
-    message: String,
-    related_information: Vec<tower_lsp::lsp_types::DiagnosticRelatedInformation>,
-}
-
-impl DiagnosticBuilder {
-    /// Create a new diagnostic builder.
-    pub fn new(message: impl Into<String>) -> Self {
-        Self {
-            range: Range::default(),
-            severity: Severity::Error,
-            source: "vize".to_string(),
-            code: None,
-            message: message.into(),
-            related_information: Vec::new(),
-        }
-    }
-
-    /// Set the range.
-    pub fn range(mut self, range: Range) -> Self {
-        self.range = range;
-        self
-    }
-
-    /// Set the severity.
-    pub fn severity(mut self, severity: Severity) -> Self {
-        self.severity = severity;
-        self
-    }
-
-    /// Set the source.
-    pub fn source(mut self, source: impl Into<String>) -> Self {
-        self.source = source.into();
-        self
-    }
-
-    /// Set the error code.
-    pub fn code(mut self, code: i32) -> Self {
-        self.code = Some(code);
-        self
-    }
-
-    /// Add related information.
-    pub fn related(
-        mut self,
-        location: tower_lsp::lsp_types::Location,
-        message: impl Into<String>,
-    ) -> Self {
-        self.related_information
-            .push(tower_lsp::lsp_types::DiagnosticRelatedInformation {
-                location,
-                message: message.into(),
-            });
-        self
-    }
-
-    /// Build the diagnostic.
-    pub fn build(self) -> Diagnostic {
-        Diagnostic {
-            range: self.range,
-            severity: Some(self.severity.into()),
-            code: self.code.map(NumberOrString::Number),
-            source: Some(self.source),
-            message: self.message,
-            related_information: if self.related_information.is_empty() {
-                None
-            } else {
-                Some(self.related_information)
-            },
-            ..Default::default()
-        }
-    }
-}
-
-/// Precomputed byte offsets of every line start in a source string.
-///
-/// Building this once and reusing it turns the per-call O(offset) scan of
-/// [`offset_to_line_col`] into a binary search over line starts plus a short
-/// UTF-16 column scan within the target line. Callers that map many offsets
-/// against the same `content` (the diagnostics collectors, the semantic-token
-/// collectors) build it once and thread it through.
-///
-/// `line_col` reproduces [`offset_to_line_col`] byte-for-byte, including the
-/// UTF-16 code-unit column semantics LSP requires.
-pub(super) struct LineIndex<'a> {
-    source: &'a str,
-    /// Byte offset of the start of each line. `line_starts[0]` is always `0`.
-    line_starts: Vec<usize>,
-}
-
-impl<'a> LineIndex<'a> {
-    /// Build a line index in a single pass over `source`.
-    pub(super) fn new(source: &'a str) -> Self {
-        let mut line_starts = Vec::with_capacity(source.len() / 32 + 1);
-        line_starts.push(0);
-        for (idx, byte) in source.bytes().enumerate() {
-            if byte == b'\n' {
-                line_starts.push(idx + 1);
-            }
-        }
-        Self {
-            source,
-            line_starts,
-        }
-    }
-
-    /// Convert a byte offset to (line, column), both 0-indexed for LSP.
-    ///
-    /// `column` is measured in UTF-16 code units, matching the editor's
-    /// coordinate system. Offsets past the end of the source are clamped,
-    /// matching the natural EOF behavior of the legacy scan.
-    pub(super) fn line_col(&self, offset: usize) -> (u32, u32) {
-        let offset = offset.min(self.source.len());
-
-        // Greatest line start <= offset.
-        let line = match self.line_starts.binary_search(&offset) {
-            Ok(line) => line,
-            Err(next) => next - 1,
-        };
-        let line_start = self.line_starts[line];
-
-        // Sum UTF-16 code units of every character on this line whose byte
-        // start is before `offset`. Bounded to the target line, and tolerant of
-        // an `offset` that falls mid-character (it counts the partial char,
-        // matching the legacy scan's break-at-top behavior) so it never panics.
-        let mut col = 0u32;
-        for (i, ch) in self.source[line_start..].char_indices() {
-            if line_start + i >= offset {
-                break;
-            }
-            col += ch.len_utf16() as u32;
-        }
-
-        (line as u32, col)
-    }
-}
-
-/// Convert byte offset to (line, column) - both 0-indexed for LSP.
-///
-/// Thin wrapper over [`LineIndex`] for single-shot callers. Hot paths that map
-/// many offsets against the same content build a [`LineIndex`] once and call
-/// [`LineIndex::line_col`] instead, so the only remaining caller is the parity
-/// unit test below.
-#[cfg(test)]
-pub(super) fn offset_to_line_col(source: &str, offset: usize) -> (u32, u32) {
-    LineIndex::new(source).line_col(offset)
-}
-
-#[cfg(test)]
-mod tests {
-    use std::fs;
-
-    use super::{DiagnosticBuilder, DiagnosticService, Severity, offset_to_line_col, sources};
-    use crate::server::ServerState;
-    use tower_lsp::lsp_types::{Diagnostic, DiagnosticSeverity, NumberOrString, Url};
-
-    fn state_with_lsp_diagnostics(lint: bool, typecheck: bool) -> ServerState {
-        let state = ServerState::new();
-        state.apply_lsp_initialization_options(Some(&serde_json::json!({
-            "lint": lint,
-            "typecheck": typecheck
-        })));
-        state
-    }
-
-    fn state_with_ecosystem_diagnostics() -> ServerState {
-        let state = ServerState::new();
-        state.apply_lsp_initialization_options(Some(&serde_json::json!({
-            "ecosystem": true
-        })));
-        state
-    }
-
-    #[test]
-    fn test_diagnostic_builder() {
-        let diagnostic = DiagnosticBuilder::new("Test error")
-            .severity(Severity::Warning)
-            .source("test")
-            .code(42)
-            .build();
-
-        assert_eq!(diagnostic.message, "Test error");
-        assert_eq!(diagnostic.severity, Some(DiagnosticSeverity::WARNING));
-        assert_eq!(diagnostic.source, Some("test".to_string()));
-        assert_eq!(diagnostic.code, Some(NumberOrString::Number(42)));
-    }
-
-    #[test]
-    fn test_severity_conversion() {
-        assert_eq!(
-            DiagnosticSeverity::from(Severity::Error),
-            DiagnosticSeverity::ERROR
-        );
-        assert_eq!(
-            DiagnosticSeverity::from(Severity::Warning),
-            DiagnosticSeverity::WARNING
-        );
-        assert_eq!(
-            DiagnosticSeverity::from(Severity::Information),
-            DiagnosticSeverity::INFORMATION
-        );
-        assert_eq!(
-            DiagnosticSeverity::from(Severity::Hint),
-            DiagnosticSeverity::HINT
-        );
-    }
-
-    #[test]
-    fn offset_to_line_col_counts_utf16_code_units() {
-        let source = "const icon = \"😀\"; missing";
-        let offset = source.find("missing").unwrap();
-
-        assert_eq!(offset_to_line_col(source, offset), (0, 19));
-    }
-
-    #[test]
-    fn collect_short_circuits_dependent_diagnostics_after_sfc_parse_error() {
-        let state = state_with_lsp_diagnostics(true, true);
-        let uri = Url::parse("file:///Broken.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            "<template><div></div>".to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-        let diagnostic_sources: Vec<_> = diagnostics
-            .iter()
-            .filter_map(|diagnostic| diagnostic.source.as_deref())
-            .collect();
-
-        assert_eq!(diagnostic_sources, vec![sources::SFC_PARSER]);
-    }
-
-    #[test]
-    fn collect_keeps_type_diagnostics_for_parseable_sfc() {
-        let state = state_with_lsp_diagnostics(false, true);
-        let uri = Url::parse("file:///Component.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            "<script setup>const props = defineProps(['count'])</script><template>{{ props.count }}</template>".to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-
-        assert!(
-            diagnostics
-                .iter()
-                .any(|diagnostic| diagnostic.source.as_deref() == Some(sources::TYPE_CHECKER))
-        );
-        assert!(
-            !diagnostics
-                .iter()
-                .any(|diagnostic| diagnostic.source.as_deref() == Some(sources::SFC_PARSER))
-        );
-    }
-
-    #[test]
-    fn collect_reports_props_destructure_default_type_mismatch_for_lsp() {
-        // Editor should surface the same DEFINE_PROPS_DESTRUCTURE_DEFAULT_TYPE
-        // error that `vize check` shows, since TypeScript's checker cannot
-        // detect a destructure default that conflicts with a Vue prop type.
-        let state = state_with_lsp_diagnostics(false, false);
-        let uri = Url::parse("file:///Bad.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            r#"<script setup lang="ts">
-const { msg = 0 } = defineProps<{ msg?: string }>();
-</script>
-
-<template>
-  <div>{{ msg }}</div>
-</template>
-"#
-            .to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-        let diagnostic = diagnostics
-            .iter()
-            .find(|diagnostic| diagnostic.source.as_deref() == Some(sources::SFC_COMPILER))
-            .expect("expected an SFC compile diagnostic");
-        assert!(
-            diagnostic
-                .message
-                .contains("DEFINE_PROPS_DESTRUCTURE_DEFAULT_TYPE"),
-            "expected DEFINE_PROPS_DESTRUCTURE_DEFAULT_TYPE in message, got: {}",
-            diagnostic.message
-        );
-        assert_eq!(diagnostic.severity, Some(DiagnosticSeverity::ERROR));
-
-        // The diagnostic must point at the offending default value (`0` on the
-        // second SFC line, after `const { msg = `), not at the start of the
-        // `<script setup>` block (line 0), which was the previous behavior.
-        assert_eq!(
-            diagnostic.range.start,
-            tower_lsp::lsp_types::Position {
-                line: 1,
-                character: 14,
-            },
-            "diagnostic should point at the default value `0`, got {:?}",
-            diagnostic.range,
-        );
-        assert_eq!(
-            diagnostic.range.end,
-            tower_lsp::lsp_types::Position {
-                line: 1,
-                character: 15,
-            },
-            "diagnostic should span the single-character default `0`, got {:?}",
-            diagnostic.range,
-        );
-    }
-
-    #[test]
-    fn collect_does_not_report_sfc_compile_diagnostic_for_valid_default() {
-        let state = state_with_lsp_diagnostics(false, false);
-        let uri = Url::parse("file:///Good.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            r#"<script setup lang="ts">
-const { msg = "ok" } = defineProps<{ msg?: string }>();
-</script>
-
-<template>
-  <div>{{ msg }}</div>
-</template>
-"#
-            .to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-        assert!(
-            diagnostics
-                .iter()
-                .all(|diagnostic| diagnostic.source.as_deref() != Some(sources::SFC_COMPILER)),
-            "expected no SFC compile diagnostics, got: {:?}",
-            diagnostics
-                .iter()
-                .map(|diagnostic| (diagnostic.source.as_deref(), diagnostic.message.as_str()))
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn collect_surfaces_sfc_level_lint_diagnostics() {
-        let state = state_with_lsp_diagnostics(true, false);
-        let uri = Url::parse("file:///OutOfOrder.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            "<template><div></div></template>\n<script setup>const count = 1</script>".to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-        let diagnostic = diagnostics
-            .iter()
-            .find(|diagnostic| {
-                diagnostic.source.as_deref() == Some(sources::LINTER)
-                    && diagnostic.code
-                        == Some(NumberOrString::String("vue/sfc-element-order".to_string()))
-            })
-            .expect("SFC-level lint diagnostic");
-
-        assert_eq!(diagnostic.range.start.line, 1);
-        assert!(diagnostic.message.contains("<script> should come before"));
-    }
-
-    #[test]
-    fn collect_lints_standalone_html_with_configured_rule() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("vize.config.json"),
-            r#"{
-                "lsp": { "lint": true },
-                "linter": {
-                    "rules": {
-                        "script/no-options-api": "error"
-                    }
-                }
-            }"#,
-        )
-        .unwrap();
-
-        let state = ServerState::new();
-        state.load_lsp_config(dir.path());
-
-        let source_path = dir.path().join("index.html");
-        let uri = Url::from_file_path(&source_path).unwrap();
-        let source = r##"<!doctype html>
-<html>
-<head>
-  <script src="https://unpkg.com/vue@3/dist/vue.global.js"></script>
-</head>
-<body>
-  <div id="app">{{ count }}</div>
-  <script>
-Vue.createApp({
-  data() {
-    return { count: 0 }
-  }
-}).mount("#app")
-  </script>
-</body>
-</html>
-"##;
-        state
-            .documents
-            .open(uri.clone(), source.to_string(), 1, "html".to_string());
-        state.update_virtual_docs(&uri, source);
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-
-        assert!(state.get_virtual_docs(&uri).is_some());
-        assert!(
-            diagnostics
-                .iter()
-                .all(|diagnostic| diagnostic.source.as_deref() != Some(sources::SFC_PARSER))
-        );
-        assert!(diagnostics.iter().any(|diagnostic| {
-            diagnostic.source.as_deref() == Some(sources::LINTER)
-                && diagnostic.code
-                    == Some(NumberOrString::String("script/no-options-api".to_string()))
-        }));
-    }
-
-    #[test]
-    fn collect_reports_unknown_file_route_params() {
-        let dir = tempfile::tempdir().unwrap();
-        let source_path = dir.path().join("src/pages/users/[id].vue");
-        fs::create_dir_all(source_path.parent().unwrap()).unwrap();
-        let source = r#"<script setup lang="ts">
-import { useRoute } from "vue-router"
-const route = useRoute()
-route.params.slug
-</script>"#;
-        fs::write(&source_path, source).unwrap();
-
-        let state = state_with_ecosystem_diagnostics();
-        let uri = Url::from_file_path(&source_path).unwrap();
-        state
-            .documents
-            .open(uri.clone(), source.to_string(), 1, "vue".to_string());
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-
-        assert!(diagnostics.iter().any(|diagnostic| {
-            diagnostic.source.as_deref() == Some("vize/ecosystem")
-                && diagnostic.code
-                    == Some(NumberOrString::String(
-                        "ecosystem/vue-router-route-param".to_string(),
-                    ))
-        }));
-    }
-
-    #[test]
-    fn collect_reports_missing_define_art_source_file() {
-        let dir = tempfile::tempdir().unwrap();
-        let source_path = dir.path().join("MissingSource.art.vue");
-        let uri = Url::from_file_path(&source_path).unwrap();
-        let source = r#"<script setup lang="ts">
-defineArt("./Missing.vue", {
-  title: "Missing",
-});
-</script>
-
-<art>
-  <variant name="Default">
-    <Missing />
-  </variant>
-</art>
-"#;
-
-        let state = state_with_lsp_diagnostics(true, false);
-        state
-            .documents
-            .open(uri.clone(), source.to_string(), 1, "art-vue".to_string());
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-
-        assert!(diagnostics.iter().any(|diagnostic| {
-            diagnostic.source.as_deref() == Some(sources::MUSEA)
-                && diagnostic.code
-                    == Some(NumberOrString::String(
-                        "musea/define-art-source-not-found".to_string(),
-                    ))
-        }));
-    }
-
-    #[test]
-    fn collect_reports_unknown_route_path_params() {
-        let source = r#"<script setup lang="ts">
-import { useRoute } from "vue-router"
-const route = useRoute()
-route.params.slug
-</script>
-<route lang="json">
-{ "path": "/users/:id" }
-</route>"#;
-
-        let state = state_with_ecosystem_diagnostics();
-        let uri = Url::parse("file:///RoutePathParams.vue").unwrap();
-        state
-            .documents
-            .open(uri.clone(), source.to_string(), 1, "vue".to_string());
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-        let diagnostic = diagnostics
-            .iter()
-            .find(|diagnostic| {
-                diagnostic.source.as_deref() == Some("vize/ecosystem")
-                    && diagnostic.code
-                        == Some(NumberOrString::String(
-                            "ecosystem/vue-router-route-param".to_string(),
-                        ))
-            })
-            .expect("unknown route path param diagnostic");
-
-        assert!(diagnostic.message.contains("Available params: id"));
-    }
-
-    #[test]
-    fn collect_reports_workspace_i18n_missing_key() {
-        let dir = tempfile::tempdir().unwrap();
-        let source_path = dir.path().join("src/components/LoginButton.vue");
-        let locale_path = dir.path().join("src/locales/en.json");
-        fs::create_dir_all(source_path.parent().unwrap()).unwrap();
-        fs::create_dir_all(locale_path.parent().unwrap()).unwrap();
-        fs::write(&locale_path, r#"{ "auth": { "login": "Log in" } }"#).unwrap();
-
-        let source = r#"<script setup lang="ts">
-const title = t("auth.missing")
-</script>"#;
-        fs::write(&source_path, source).unwrap();
-
-        let state = state_with_ecosystem_diagnostics();
-        let uri = Url::from_file_path(&source_path).unwrap();
-        state
-            .documents
-            .open(uri.clone(), source.to_string(), 1, "vue".to_string());
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-
-        assert!(diagnostics.iter().any(|diagnostic| {
-            diagnostic.source.as_deref() == Some("vize/ecosystem")
-                && diagnostic.code
-                    == Some(NumberOrString::String(
-                        "ecosystem/vue-i18n-no-missing-key".to_string(),
-                    ))
-        }));
-    }
-
-    #[test]
-    fn collect_short_circuits_dependent_diagnostics_after_script_parse_error() {
-        let state = state_with_lsp_diagnostics(true, true);
-        let uri = Url::parse("file:///BrokenScript.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            "<script setup lang=\"ts\">\nconst count =</script>\n<template>{{ count }}</template>"
-                .to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-        let diagnostic_sources: Vec<_> = diagnostics
-            .iter()
-            .filter_map(|diagnostic| diagnostic.source.as_deref())
-            .collect();
-
-        assert_eq!(diagnostic_sources, vec![sources::SCRIPT_PARSER]);
-        assert_eq!(
-            diagnostics[0].code,
-            Some(NumberOrString::String("script-parse-error".to_string()))
-        );
-        assert_eq!(diagnostics[0].range.start.line, 1);
-    }
-
-    #[test]
-    fn collect_accepts_jsx_script_without_false_parse_diagnostic() {
-        let state = state_with_lsp_diagnostics(false, false);
-        let uri = Url::parse("file:///JsxScript.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            "<script setup lang=\"jsx\">\nconst count = 1\nconst vnode = <button>{count}</button>\n</script>"
-                .to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-
-        assert!(
-            !diagnostics
-                .iter()
-                .any(|diagnostic| diagnostic.source.as_deref() == Some(sources::SCRIPT_PARSER))
-        );
-    }
-
-    #[test]
-    fn collect_accepts_tsx_script_without_false_parse_diagnostic() {
-        let state = state_with_lsp_diagnostics(false, false);
-        let uri = Url::parse("file:///TsxScript.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            "<script setup lang=\"tsx\">\nconst count = 1\nconst vnode = <button>{count}</button>\n</script>"
-                .to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-
-        assert!(
-            !diagnostics
-                .iter()
-                .any(|diagnostic| diagnostic.source.as_deref() == Some(sources::SCRIPT_PARSER))
-        );
-    }
-
-    #[test]
-    fn collect_skips_unsupported_script_language() {
-        let state = state_with_lsp_diagnostics(false, false);
-        let uri = Url::parse("file:///CoffeeScript.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            "<script setup lang=\"coffee\">\ncount = ->\n</script>".to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-
-        assert!(
-            !diagnostics
-                .iter()
-                .any(|diagnostic| diagnostic.source.as_deref() == Some(sources::SCRIPT_PARSER))
-        );
-    }
-
-    fn lint_subset(diagnostics: &[Diagnostic]) -> Vec<Diagnostic> {
-        diagnostics
-            .iter()
-            .filter(|d| {
-                matches!(
-                    d.source.as_deref(),
-                    Some(sources::LINTER) | Some(sources::MUSEA)
-                )
-            })
-            .cloned()
-            .collect()
-    }
-
-    #[test]
-    fn collect_lint_only_equals_lint_subset_of_collect() {
-        // With both lint and typecheck enabled, collect() produces LINTER and
-        // TYPE_CHECKER diagnostics; collect_lint_only() must return exactly the
-        // LINTER/MUSEA subset (same diagnostics, same order) and nothing else.
-        let state = state_with_lsp_diagnostics(true, true);
-        let uri = Url::parse("file:///OutOfOrder.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            "<template><div></div></template>\n<script setup>const count = 1</script>".to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let full = DiagnosticService::collect(&state, &uri);
-        let lint_only = DiagnosticService::collect_lint_only(&state, &uri);
-
-        // Sanity: the full pipeline really did surface a lint diagnostic here.
-        assert!(
-            full.iter()
-                .any(|d| d.source.as_deref() == Some(sources::LINTER)),
-            "expected a lint diagnostic from the full pipeline"
-        );
-        // The lint-only result is byte-for-byte the lint/musea subset.
-        assert_eq!(lint_only, lint_subset(&full));
-        // And it carries no non-lint sources (no type-check / parser / sfc-compile).
-        assert!(
-            lint_only.iter().all(|d| {
-                matches!(
-                    d.source.as_deref(),
-                    Some(sources::LINTER) | Some(sources::MUSEA)
-                )
-            }),
-            "collect_lint_only must only return lint/musea diagnostics, got: {:?}",
-            lint_only
-                .iter()
-                .map(|d| d.source.as_deref())
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn collect_lint_only_short_circuits_on_parse_error() {
-        // A block parse error gates lint in the full pipeline, so the lint-only
-        // path must also return nothing — matching what hover would otherwise see.
-        let state = state_with_lsp_diagnostics(true, true);
-        let uri = Url::parse("file:///BrokenTemplateLintOnly.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            "<script setup lang=\"ts\">const count = 1</script>\n<template><div>{{ count }}</template>"
-                .to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let full = DiagnosticService::collect(&state, &uri);
-        let lint_only = DiagnosticService::collect_lint_only(&state, &uri);
-
-        assert_eq!(lint_only, lint_subset(&full));
-        assert!(lint_only.is_empty());
-    }
-
-    #[test]
-    fn collect_lint_only_is_empty_when_lint_disabled() {
-        // Hover gates on is_lsp_lint_enabled, but the collector is defensively
-        // empty when lint is off even if typecheck is on.
-        let state = state_with_lsp_diagnostics(false, true);
-        let uri = Url::parse("file:///NoLint.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            "<script setup>const props = defineProps(['count'])</script><template>{{ props.count }}</template>".to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        assert!(DiagnosticService::collect_lint_only(&state, &uri).is_empty());
-    }
-
-    #[test]
-    fn collect_short_circuits_dependent_diagnostics_after_template_parse_error() {
-        let state = state_with_lsp_diagnostics(true, true);
-        let uri = Url::parse("file:///BrokenTemplate.vue").unwrap();
-        state.documents.open(
-            uri.clone(),
-            "<script setup lang=\"ts\">const count = 1</script>\n<template><div>{{ count }}</template>"
-                .to_string(),
-            1,
-            "vue".to_string(),
-        );
-
-        let diagnostics = DiagnosticService::collect(&state, &uri);
-        let diagnostic_sources: Vec<_> = diagnostics
-            .iter()
-            .filter_map(|diagnostic| diagnostic.source.as_deref())
-            .collect();
-
-        assert_eq!(diagnostic_sources, vec![sources::TEMPLATE_PARSER]);
-    }
 }
